+++ manylinux/install_requirements.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 set -e
 
-for d in /opt/python/cp{27,34,35,36,37}*; do
+for d in /opt/python/cp{27,35,36,37}*; do
     $d/bin/pip install -r /tmp/spead2/requirements.txt
 done